+++ bioblend/_tests/test_util.py
@@ -7,23 +7,12 @@ import unittest
 
 import bioblend.galaxy
 
-NO_CLOUDMAN_MESSAGE = "CloudMan required and no CloudMan AMI configured."
 NO_GALAXY_MESSAGE = "Externally configured Galaxy required, but not found. Set BIOBLEND_GALAXY_URL and BIOBLEND_GALAXY_API_KEY to run this test."
 NO_TOOLSHED_MESSAGE = (
     "Externally configured ToolShed required, but not found. Set BIOBLEND_TOOLSHED_URL to run this test."
 )
 OLD_GALAXY_RELEASE = "Testing on Galaxy %s, but need %s to run this test."
 MISSING_TOOL_MESSAGE = "Externally configured Galaxy instance requires tool %s to run test."
-
-
-def skip_unless_cloudman():
-    """Decorate tests with this to skip the test if CloudMan is not
-    configured.
-    """
-    if "BIOBLEND_AMI_ID" not in os.environ:
-        return unittest.skip(NO_CLOUDMAN_MESSAGE)
-    else:
-        return lambda f: f
 
 
 def skip_unless_toolshed():