+++ bioblend/util/__init__.py
@@ -8,24 +8,6 @@ from typing import (
 )
 
 
-class Bunch:
-    """
-    A convenience class to allow dict keys to be represented as object fields.
-
-    The end result is that this allows a dict to be to be represented the same
-    as a database class, thus the two become interchangeable as a data source.
-    """
-
-    def __init__(self, **kwargs: Any) -> None:
-        self.__dict__.update(kwargs)
-
-    def __repr__(self) -> str:
-        """
-        Return the contents of the dict in a printable representation
-        """
-        return str(self.__dict__)
-
-
 class FileStream(NamedTuple):
     name: str
     fd: IO
@@ -75,6 +57,6 @@ def abstractclass(decorated_cls: Type[T]) -> Type[T]:
 
 
 __all__ = (
-    "Bunch",
+    "abstractclass",
     "attach_file",
 )